+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-7286] Added audit UploadJobs functions
+
 ## [7.14.1] 2025-12-09
 
 - [TD-7678] Update TdBg functions for search concepts by name, domain and template